+++ pytest_reportportal/plugin.py
@@ -598,6 +598,11 @@ def pytest_addoption(parser) -> None:
         default=str(MAX_LOG_BATCH_PAYLOAD_SIZE),
         help="Maximum payload size in bytes of async batch log requests",
     )
+    parser.addini(
+        "rp_log_batch_payload_size",
+        default=str(MAX_LOG_BATCH_PAYLOAD_SIZE),
+        help="DEPRECATED: Maximum payload size in bytes of async batch log requests",
+    )
     parser.addini("rp_ignore_attributes", type="args", help="Ignore specified pytest markers, i.e parametrize")
     parser.addini(
         "rp_is_skipped_an_issue", default=True, type="bool", help="Treat skipped tests as required investigation"