+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Return back deprecated `rp_log_batch_payload_size` parameter for sake of backward compatibility, by @HardNorth
 
 ## [5.5.3]
 ### Added